Moved the net folder into the newly created dflowfm folder and updated the references.
