+++ hydrolib/core/io/dflowfm/net/reader.py
@@ -12,7 +12,7 @@ import numpy as np
 from hydrolib.core.basemodel import BaseModel
 
 if TYPE_CHECKING:
-    from .models import Link1d2d, Mesh1d, Mesh2d
+    from ..dflowfm.net.models import Link1d2d, Mesh1d, Mesh2d
 
 
 class UgridReader:
